+++ src/Resources/contao/dca/tl_content.php
@@ -14,6 +14,7 @@ declare(strict_types=1);
 
 use Fiedsch\LigaverwaltungBundle\Helper\DCAHelper;
 use Fiedsch\LigaverwaltungBundle\Model\HighlightModel;
+//use Contao\DataContainer;
 
 /* Ligenliste */
 $GLOBALS['TL_DCA']['tl_content']['palettes']['ligenliste'] = '{type_legend},type,headline;{auswahl_legend},verband,saison;{template_legend:hide},customTpl;{protected_legend:hide},protected;{expert_legend:hide},guests,cssID,space;{invisible_legend:hide},invisible,start,stop';
@@ -145,7 +146,7 @@ $GLOBALS['TL_DCA']['tl_content']['fields']['ligen'] = [
     'inputType' => 'checkboxWizard',
     'filter' => false,
     'sorting' => false,
-    // 'flag'        => 1, // Sort by initial letter ascending
+    //'flag'        => DataContainer::SORT_INITIAL_LETTER_ASC,
     'relation' => ['type' => 'belongsTo', 'load' => 'lazy'],
     'foreignKey' => 'tl_liga.name',
     'eval' => ['multiple' => true, 'tl_class' => 'w50 clr'],