+++ src/Resources/contao/dca/tl_spielort.php
@@ -12,6 +12,8 @@ declare(strict_types=1);
  * @license https://opensource.org/licenses/MIT
  */
 
+use Contao\DataContainer;
+
 \Contao\System::loadLanguageFile('default');
 
 $GLOBALS['TL_DCA']['tl_spielort'] = [
@@ -28,8 +30,8 @@ $GLOBALS['TL_DCA']['tl_spielort'] = [
 
     'list' => [
         'sorting' => [
-            'mode' => 1, // Records are sorted by a fixed field
-            'flag' => 1, // Sort by initial letter ascending
+            'mode' => DataContainer::MODE_SORTED,
+            'flag' => DataContainer::SORT_INITIAL_LETTER_ASC,
             'fields' => ['name'],
             'panelLayout' => 'sort,filter;search,limit',
         ],
@@ -86,7 +88,7 @@ $GLOBALS['TL_DCA']['tl_spielort'] = [
             'sorting' => true,
             'search' => true,
             'exclude' => true,
-            'flag' => 1, // sort by initial letter ascending
+            'flag' => DataContainer::SORT_INITIAL_LETTER_ASC,
             'inputType' => 'text',
             'eval' => ['maxlength' => 128, 'tl_class' => 'w50'],
             'sql' => "varchar(128) NOT NULL default ''",